+++ src/utils/logger.ts
@@ -1,49 +1,50 @@
+/* eslint-disable no-console */
 import chalk from 'chalk'
 import figures from 'figures'
-import { createLogger, format, transports } from 'winston'
 
 import { logLevels } from './logger.constants'
-import { ILogger, ILoggerFormat } from './logger.interface'
 
 export class Logger {
-  static readonly levels = {
-    [logLevels.fail]: 1,
-    [logLevels.skip]: 2,
-    [logLevels.success]: 3,
-    [logLevels.data]: 4,
-    [logLevels.start]: 5
+
+  public fail (message: string): void {
+    message = this.parseMessage(logLevels.fail, message)
+    console.error(message)
   }
 
-  public log: ILogger
-
-  constructor (logLevel?: logLevels) {
-    this.log = this.initiateLogger(logLevel)
+  public skip (message: string): void {
+    message = this.parseMessage(logLevels.fail, message)
+    console.warn(message)
   }
 
-  private initiateLogger (logLevel?: logLevels): ILogger {
-    const loglevel: string = logLevel
-    const logFormat = format.printf(({ level, message }: ILoggerFormat) => {
-      // parse multi line messages
-      let multiLineMessage = message?.split('\n')
-      multiLineMessage = multiLineMessage?.map((msg) => {
-        // format messages
-        return this.logColoring({ level, message: msg })
-      })
-      // join back multi line messages
-      message = multiLineMessage?.join('\n')
-      return message
+  public success (message: string): void {
+    message = this.parseMessage(logLevels.success, message)
+    console.log(message)
+  }
+
+  public data (message: string): void {
+    message = this.parseMessage(logLevels.data, message)
+    console.info(message)
+  }
+
+  public start (message: string): void {
+    message = this.parseMessage(logLevels.start, message)
+    console.info(message)
+  }
+
+  private parseMessage (level: logLevels ,message: string): string {
+    // parse multi line messages
+    let multiLineMessage = message?.split('\n')
+    multiLineMessage = multiLineMessage?.map((msg) => {
+      // format messages
+      return this.logColoring({ level, message: msg })
     })
+    // join back multi line messages
+    message = multiLineMessage?.join('\n')
 
-    return createLogger({
-      level: loglevel || logLevels.start,
-      silent: (loglevel === logLevels.silent),
-      format: format.combine(logFormat),
-      levels: Logger.levels,
-      transports: [new transports.Console()]
-    }) as ILogger
+    return message
   }
 
-  private logColoring ({ level, message }: ILoggerFormat): string {
+  private logColoring ({ level, message }: {level: logLevels, message: string}): string {
     let icon: string
 
     // do the coloring
@@ -74,6 +75,6 @@ export class Logger {
       break
     }
 
-    return coloring(`${icon}  ${message}`)
+    return coloring(`${icon} ${message}`)
   }
 }